DB-5692: Refactor cms-kit bundle
