+++ .changeset/quick-dancers-cover.md
@@ -1,0 +1,5 @@
+---
+'@pantheon-systems/cms-kit': patch
+---
+
+Refactored bundle. There should be no changes to consumers of this package
