+++ packages/cms-kit/.eslintrc
@@ -1,6 +1,7 @@
 {
-  "extends": "@pantheon-systems/eslint-config",
-  "parserOptions": {
-    "project": ["./tsconfig.json"]
-  }
+	"extends": "@pantheon-systems/eslint-config",
+	"parserOptions": {
+		"project": ["./tsconfig.json"]
+	},
+	"ignorePatterns": ["tsup.config.ts"]
 }
