+++ packages/cms-kit/src/index.ts
@@ -1,0 +1,1 @@
+export { setSurrogateKeyHeader } from './lib/setSurrogateKeyHeader.js';
