--- packages/cms-kit/src/utils/index.ts
@@ -1,1 +1,0 @@
-export { default as setSurrogateKeyHeader } from './setSurrogateKeyHeader';
