+++ packages/cms-kit/tsconfig.json
@@ -1,8 +1,8 @@
 {
 	"extends": "../../tsconfig.json",
 	"compilerOptions": {
+		"types": ["vitest/globals"],
 		"declaration": true,
 		"declarationDir": "./dist"
-	},
-	"exclude": ["__tests__"]
+	}
 }
